Fix pointer handling in busybox's vi
The upstream code uses realloc to increase an existing memory
region and then adds the pointer difference (bias) between the
old and the new pointer the old pointer to obtain the pointer
in the reallocated region. With CHERI this fails because the
resulting pointer is out of bounds with respect to the origin
allocation. The situation is complicated by the fact that the
calculated bias is returned from the function that calculates
it and callers use it to fixup their pointers.

Fix fix this with minimal changes to uptream keep the pointer
calculations done by upstream but post-process the resulting
pointers. Post processing calculates the offset of the pointer
within the reallocated region and adds this offset to the base
of the region thus using pointer bounds of the new regions
without any change to the address.