vi: fix cheri crash when opening an empty file
touch ./empty && vi ./empty

crashes vi on a cheri system.

The count_lines function gets two pointers for the upper and lower bounds
and returns the number of lines in this range. The two pointers should
point to the editor buffer (between the global "text" and "end" pointers).

count_lines seems to be optimized for speed and has no range checks for
its parameters. Some callers set one of the bounds to the byte before the
edit buffer. On cheri systems, this causes a bounds violation when this
byte is read.

This fault happens at startup when the buffer is empty and on other
occassions. Add a simple check to count_lines to fix it. The overhead
in count_lines and the increase in code size should be neglectable.