vi: fix oob access by the X command.
X deletes the character left of the cursor.

Generally, dot points to the cursor position, text points to the buffer
that contains the whole file.

For the X command, dot[-1] is deleted. Add a check to prevent an
out-of-bounds access when the cursor is at the beginning of the file
(dot == text). On a cheri system, the bounds violation triggers a cheri
exception.