cheri: propagate bufsiz1 alignment info
Most applets use a struct of global variables that's stored in the bufsiz1 array.
These structs should be capability-size aligned to allow for efficient access to
their components. Otherwise, cherillvm produces warnings such as

   networking/ping.c:960:15: warning: found underaligned store of capability
   type (aligned to 8 bytes instead of 16). Will use memcpy() instead of
   capability load to preserve tags if it is aligned correctly at runtime
   [-Wcheri-inefficient]

busybox goes to great lengths to place bufsiz1 into unused memory at the
end of the bss segment. This involves multiple linker runs and a script
to generate the bufsiz1 definition. bufsiz1 is aligned to sizeof(long
long) but this alignment info isn't propagated to all users of bufsiz1.

This patch changes the alignment to capability size, which might be
larger than sizeof(long long), and propagates alignment info to the
callers. This eliminates hundreds of -Wcheri-inefficient warnings.