cheri: fix cast for PR_GET_NAME argument
According to prctl(2), the function is defined as

int prctl(int option, unsigned long arg2,
   unsigned long arg3, unsigned long arg4, unsigned long arg5);

but the manpage clarifies that for PR_SET_NAME, arg2 is cast to (char *).

busybox passes (long)comm as arg2 where comm is a char *. Replace the
cast with uintptr_t to preserve the capability on a cheri system. This
should be safe for non-cheri as well.