vi: use correct base for F after text buffer resize
printf 'a%.s' `seq 1 10300` | vi -

produces a cheri fault.

If no file is loaded at startup, vi reserves an initial text buffer of
10240 bytes. realloc is called to increase this buffer as needed.

After a realloc, all pointers to locations inside the previous buffer are
updated to point to the new buffer. This is done by adding the offset
between the old and new buffers. For cheri, the F macro was introduced to
fix up the capabilities after adding the offset. F recreates the
capabilities with the bounds of the text buffer ("text" points to this
buffer).

When the text buffer itself is moved by the realloc, we have to set
"text" to the new location before calling F. Otherwise, we'll use the
bounds of the old text buffer.

Fixes: e3fa0b3d1 ("Fix pointer handling in busybox's vi")