vi: fix oob-accesses for 'B' command on file start
In vi, the command 'B' allows to jump to the beginning of the previous
word. Due to insufficient checks, its handling can causes currently two
out-of-bounds memory accesses.

Generally, vi allocates a buffer for the text hold by the current editor
session. The char pointer 'text' points to the start of the buffer, the
char pointer 'dot' is a pointer to the current cursor location within
the allocated buffer.

When handling the 'B' command (together with 'E' and 'W') in 'do_cmd',
the condition 'isspace(dot[dir])' is checked , where 'dir == -1' for
'B'. If 'dot' points to the start of the text buffer, so 'dot == text',
this accesses memory out of bounds. Add a check to make sure that the
relative access of 'dot' is not before the start of the buffer.

During further handling of the 'B' command, a similar out of bounds
access happens during calculation of the new 'dot'. The function
'skip_thing' invokes 'st_test' with a current pointer 'p' and a
direction 'dir'. In the same way, 'p[inc]' is accessed without
previously checking whether this memory is inside the allocated text
buffer. Move the bounds check before the invocation of 'st_test'.

On CHERI systems with fine-grained memory protection through
architectural capabilities, these out of bounds accesses caused bounds
violations that allowed detecting these bugs.

function                                             old     new   delta
do_cmd                                              8376    8536    +160
------------------------------------------------------------------------------
(add/remove: 0/0 grow/shrink: 1/0 up/down: 160/0)             Total: 160 bytes

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>